(net/build) suppress some warnings (gcc/bsd) and add sqlite3 to 'client-only' build deps
FossilOrigin-Name: 53b1d78e2420a4513a5a8e05aa2887e57e76ca8b41a9d83e7fe743b6240362c3

+++ src/a12/net/dir_lua.c
@@ -35,7 +35,7 @@ static void dump_stack(lua_State* L, FILE* dst)
 	int top = lua_gettop(L);
 	fprintf(dst, "-- stack dump (%d)--\n", top);
 
-	for (size_t i = 1; i <= top; i++){
+	for (int i = 1; i <= top; i++){
 		int t = lua_type(L, i);
 
 		switch (t){
